// src/timesheet-web/src/app/api/models/week-backlog-group.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { WeekRow } from '../models/week-row';
export interface WeekBacklogGroup {
  assigneeName?: string | null;
  backlogCode?: string | null;
  backlogId?: number;
  periodMonth?: string | null;
  project?: string | null;
  tasks?: Array<WeekRow> | null;
  type?: string | null;
}